requirements.txt
flask
firebase-admin
google-generativeai
gunicorn
razorpay
